fix: change redemption_log.created_at to BigInteger
밀리초 타임스탬프가 Integer(32비트) 범위 초과
BigInteger(64비트)로 변경 + DB 마이그레이션 추가

+++ app/models/redemption_log.py
@@ -1,7 +1,7 @@
 """
 Redemption Log model for RiderVoice AI Backend
 """
-from sqlalchemy import Column, Integer, String, Boolean, Text
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, Text
 from app.database import Base
 
 
@@ -15,7 +15,7 @@ class RedemptionLog(Base):
     success = Column(Boolean, nullable=False)
     error_message = Column(Text, nullable=True)
     ip_address = Column(String(50), nullable=True)
-    created_at = Column(Integer, nullable=False)
+    created_at = Column(BigInteger, nullable=False)
 
     def __init__(self, **kwargs):
         import time
